requirements.txt
# ========================================
# RASA + PHOBERT COMPATIBLE VERSIONS
# Python 3.10 required
# Tested and working configuration
# ========================================

# Core Rasa Framework
rasa==3.6.20
rasa-sdk==3.6.2

# NumPy - CRITICAL: Must be < 1.24 for TensorFlow 2.12
numpy==1.23.5

# Transformers ecosystem for PhoBERT
transformers==4.35.2
torch==2.1.2
tokenizers==0.15.0
sentencepiece==0.1.99
huggingface_hub==0.25.2
PySide6==6.7.2
faiss-cpu==1.8.0.post1
openai==1.48.0
groq==0.9.0
google-generativeai==0.8.3

# Data processing - downgrade pandas
pandas==1.5.3

# Web framework
flask==3.0.3
flask-cors==5.0.0
requests==2.32.3

# NLP (SpaCy sẽ được Rasa cài tự động)
# spacy==3.7.6  # Không cần khai báo riêng

# Utilities
python-dotenv==1.0.1
rich==13.8.1
tqdm==4.66.5

# Code quality
black==22.12.0
isort==5.13.2
ruff==0.6.5

# Development & Notebooks
jupyter==1.1.1
ipykernel==6.29.5

# Visualization - downgrade matplotlib
matplotlib==3.5.3
seaborn==0.12.2
plotly==5.24.0

# Production server
gunicorn==23.0.0

# Database
sqlalchemy==1.4.52

# JSON schema validation
fastjsonschema==2.20.0

# Text preprocessing for Vietnamese
regex==2024.5.15